+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - docs: Add link to related github issue for `slurmrestd` TCP/IP socket
   limitation in architecture page.
 
+### Fixed
+- backend: handle rfl.settings.errors.SettingsSiteLoaderError when loading
+  gateway and backend site configuration (#317).
+
 ## [3.1.0] - 2024-07-03
 
 ### Added